refactor(skills): organize runtime code into dedicated packages
