+++ config/constants/exports.py
@@ -15,6 +15,9 @@ EXPORTS: dict[str, str] = {
     "CI_REPAIR_REPORT_BUILDER": "ci_repair",
     "CI_REPAIR_CRON": "ci_repair",
     "ONBOARDING_SKILL_NAME": "skills",
+    "SKILL_FILENAME": "skills",
+    "SKILL_REPORT_SUFFIX": "skills",
+    "SKILLS_HEADER": "skills",
     "ANALYZING_GITHUB_CI_PERFORMANCE_SKILL_NAME": "skills",
     "SCHEDULING_GITHUB_CI_FIXES_SKILL_NAME": "skills",
     "CONNECTING_SLACK_SKILL_NAME": "skills",
