+++ core/agent_harness/__init__.py
@@ -16,8 +16,8 @@ from __future__ import annotations
 
 from core.agent_harness.harness import AgentSession, SessionConfig
 from core.agent_harness.ports import OutputSink
-from core.agent_harness.prompts.skills.naming import is_legacy_skill_name, normalize_skill_name
-from core.agent_harness.prompts.skills.schedule import (
+from core.agent_harness.prompts.skills import is_legacy_skill_name, normalize_skill_name
+from core.agent_harness.prompts.skills.scheduling import (
     is_recurring_skill,
     pin_recurring_skill,
     resolve_scheduled_skill,
