+++ core/agent_harness/prompts/skills/catalog/__init__.py
@@ -1,0 +1,3 @@
+"""Skill catalog components."""
+
+__all__: list[str] = []
