+++ core/agent_harness/prompts/skills/catalog/contracts.py
