+++ core/agent_harness/prompts/skills/catalog/naming.py
