+++ core/agent_harness/prompts/skills/catalog/schema.py
