+++ core/agent_harness/prompts/skills/content/__init__.py
@@ -1,0 +1,3 @@
+"""Skill content components."""
+
+__all__: list[str] = []
