+++ core/agent_harness/prompts/skills/scheduling/__init__.py
@@ -1,0 +1,21 @@
+"""Revision-pinned execution contracts for recurring skills."""
+
+from core.agent_harness.prompts.skills.catalog.registry import find_action_skill
+from core.agent_harness.prompts.skills.scheduling.recurring import (
+    ScheduledSkillResolution,
+    is_recurring_skill,
+    pin_recurring_skill,
+    resolve_scheduled_skill,
+    skill_revision,
+    validate_skill_inputs,
+)
+
+__all__ = [
+    "ScheduledSkillResolution",
+    "find_action_skill",
+    "is_recurring_skill",
+    "pin_recurring_skill",
+    "resolve_scheduled_skill",
+    "skill_revision",
+    "validate_skill_inputs",
+]
