+++ tests/core/agent/orchestration/test_action_prompt.py
@@ -15,16 +15,13 @@ from core.agent_harness.prompts import (
     repository_context_block,
 )
 from core.agent_harness.prompts.memory.conversation import NO_HISTORY_PLACEHOLDER
-from core.agent_harness.prompts.skills.loader import (
+from core.agent_harness.prompts.skills import (
     SKILLS_HEADER,
+    clear_skills_caches,
     list_action_skills,
     load_skill_body,
-    load_skills_block,
     load_skills_index,
     skills_dir,
-)
-from core.agent_harness.prompts.skills.loader import (
-    load_skills_block as cached_load_skills_block,
 )
 from core.agent_harness.turns.turn_snapshot import TurnSnapshot
 from tests.utils.skill_cards import skill_card
@@ -144,7 +141,7 @@ def test_system_prompt_slack_fragment_documents_invented_command_example() -> No
 
 def test_morning_report_skill_closes_with_schedule_offer() -> None:
     """A run-once morning report without an offer cannot drive repeat usage."""
-    load_skills_block.cache_clear()
+    clear_skills_caches()
     body = _skill_instruction_text("delivering-morning-briefings")
     assert "propose_scheduled_delivery" in body
     assert "recurring_skill" in body
@@ -202,8 +199,10 @@ def test_skill_body_appends_sibling_report_template_but_index_stays_thin(
         encoding="utf-8",
     )
     (skill_dir / "auditing_report.md").write_text("### Findings by severity\n", encoding="utf-8")
-    monkeypatch.setattr("core.agent_harness.prompts.skills.loader.skills_dir", lambda: tmp_path)
-    cached_load_skills_block.cache_clear()
+    monkeypatch.setattr(
+        "core.agent_harness.prompts.skills.content.files.skills_dir", lambda: tmp_path
+    )
+    clear_skills_caches()
     try:
         index = load_skills_index()
         assert "auditing" in index
@@ -217,11 +216,11 @@ def test_skill_body_appends_sibling_report_template_but_index_stays_thin(
         assert f"REPORT TEMPLATE from `{report_path}`" in body
         assert body.endswith("### Findings by severity")
     finally:
-        cached_load_skills_block.cache_clear()
+        clear_skills_caches()
 
 
 def test_skills_loader_bundles_github_security_fix_skill() -> None:
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
     skill = skills_dir() / "fixing-github-security-alerts" / "SKILL.md"
     assert skill.is_file()
 
@@ -241,11 +240,11 @@ def test_skills_loader_bundles_github_security_fix_skill() -> None:
     assert "output exactly that text and stop" in body
     assert "reply in one short line" in body
     assert 'Do not say "next steps"' in body
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
 
 
 def test_skills_loader_bundles_github_ci_fix_skill() -> None:
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
     skill = skills_dir() / "fixing-github-ci" / "SKILL.md"
     assert skill.is_file()
 
@@ -257,11 +256,11 @@ def test_skills_loader_bundles_github_ci_fix_skill() -> None:
     assert "separate linked git" in body
     assert "worktree, commits on a fresh" in body
     assert 'branch="main"' in body
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
 
 
 def test_skill_matches_take_priority_over_generic_docs_answer() -> None:
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
 
     index = load_skills_index()
     body = load_skill_body(ONBOARDING_SKILL_NAME)
@@ -273,7 +272,7 @@ def test_skill_matches_take_priority_over_generic_docs_answer() -> None:
     # Skills index still rides the assembled prompt after the markdown base.
     assert SKILLS_HEADER in prompt
     assert ONBOARDING_SKILL_NAME in prompt
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
 
 
 def test_action_system_prompt_includes_context_blocks() -> None:
@@ -292,7 +291,7 @@ def test_action_system_prompt_includes_context_blocks() -> None:
 
 
 def test_skills_index_is_thin_relative_to_full_bodies() -> None:
-    cached_load_skills_block.cache_clear()
+    clear_skills_caches()
     index = load_skills_index()
     bodies = "".join(load_skill_body(skill.name) for skill in list_action_skills())
     assert index.startswith(SKILLS_HEADER)
@@ -407,7 +406,7 @@ def test_scheduling_is_never_offered_without_asking_first() -> None:
     longer inlines cron routing).
     """
     # Arrange
-    load_skills_block.cache_clear()
+    clear_skills_caches()
     skill = _skill_instruction_text("delivering-morning-briefings")
 
     # Assert — structured propose tool; creation waits on confirm / yes
