+++ tests/core/agent/prompts/test_skills_demo.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 
-import core.agent_harness.prompts.skills.loader as loader
+import core.agent_harness.prompts.skills as skills
 from config.constants.skills import ONBOARDING_SKILL_NAME, SKIP_DEMO_OPTION
 from core.agent_harness.prompts.action import build_action_system_prompt
 from core.agent_harness.prompts.action.assemble import build_action_system_prompt_envelope
@@ -24,7 +24,7 @@ from tests.utils.skill_cards import skill_card
 
 def test_child_directories_are_letter_prefixed_skill_names() -> None:
     """``<letter>-<name>/`` keeps disk order, menu order, and ``name`` from drifting apart."""
-    loader.clear_skills_caches()
+    skills.clear_skills_caches()
     for skill in getting_started_skills():
         directory = skill.path.parent.name
         letter, _, suffix = directory.partition("-")
@@ -34,7 +34,7 @@ def test_child_directories_are_letter_prefixed_skill_names() -> None:
 
 
 def test_master_menu_matches_three_unique_children_and_preserves_specialists() -> None:
-    loader.clear_skills_caches()
+    skills.clear_skills_caches()
     children = getting_started_skills()
     assert [s.name for s in children] == [
         "analyzing-github-ci-performance",
@@ -47,8 +47,8 @@ def test_master_menu_matches_three_unique_children_and_preserves_specialists() -
         "Set up an agent that improves CI/CD reliability over time",
         "Connect OpenSRE to Slack and hand off DevOps chores for your team",
     )
-    master = loader.load_skill_body(ONBOARDING_SKILL_NAME)
-    master_skill = next(s for s in loader.list_action_skills() if s.name == ONBOARDING_SKILL_NAME)
+    master = skills.load_skill_body(ONBOARDING_SKILL_NAME)
+    master_skill = next(s for s in skills.list_action_skills() if s.name == ONBOARDING_SKILL_NAME)
     # The menu is data the host runs on entry, not prose the model replays; its
     # options are the children's own labels so the two cannot drift apart.
     assert [call.tool for call in master_skill.pre_execute] == ["ask_user_choice"]
@@ -60,12 +60,12 @@ def test_master_menu_matches_three_unique_children_and_preserves_specialists() -
     for skill in children:
         assert f'skill_view(name="{skill.name}")' in master
         assert skill.path.parent.parent.name == ONBOARDING_SKILL_NAME
-        assert loader.load_skill_body(skill.name)
+        assert skills.load_skill_body(skill.name)
     analytics = next(s for s in children if s.name == "analyzing-github-ci-performance")
     # The analytics card runs its menus from the plan the model follows, not
     # from host hooks, and keeps the full tool catalog.
     assert analytics.pre_execute == ()
-    body = loader.load_skill_body("analyzing-github-ci-performance")
+    body = skills.load_skill_body("analyzing-github-ci-performance")
     assert "`Which repository should I analyze?`" in body
     assert "`What would you like to do next?`" in body
     for option in ("- Schedule local loops", "- Slack setup", "- Finish"):
@@ -80,7 +80,7 @@ def test_master_menu_matches_three_unique_children_and_preserves_specialists() -
     assert "Compare these numbers" not in body
     assert "Output its `headline`" not in body
     assert "same-day snapshot" not in body
-    fix_loop = loader.load_skill_body("scheduling-github-ci-fixes")
+    fix_loop = skills.load_skill_body("scheduling-github-ci-fixes")
     # The fix loop repairs red pull requests; it is not the analytics report
     # loop, so it never reaches for the analytics or report-scheduling tools.
     assert "schedule_ci_repair_loop" in fix_loop
@@ -91,14 +91,14 @@ def test_master_menu_matches_three_unique_children_and_preserves_specialists() -
     assert "ask_user_choice" in fix_loop
     assert menu["allow_custom"] is False
     assert GETTING_STARTED_CUSTOM not in master
-    assert loader.load_skill_body("delegating-github-ci-fixes") == ""
-    discovered = [
-        loader.validate_skill_file(path) for path in loader._iter_skill_paths(loader.skills_dir())
-    ]
+    assert skills.load_skill_body("delegating-github-ci-fixes") == ""
+    catalog = skills.read_skill_catalog()
+    assert catalog.diagnostics == ()
+    discovered = catalog.skills
     names = [skill.name for skill in discovered if skill is not None]
     assert len(names) == len(set(names))
-    assert ONBOARDING_SKILL_NAME in loader.load_skills_index()
-    assert loader.load_skill_body("fixing-github-ci")
+    assert ONBOARDING_SKILL_NAME in skills.load_skills_index()
+    assert skills.load_skill_body("fixing-github-ci")
 
 
 def test_multi_step_skills_track_progress_with_update_plan_not_step_headers() -> None:
@@ -108,7 +108,7 @@ def test_multi_step_skills_track_progress_with_update_plan_not_step_headers() ->
     contradicted the base prompt's header rules and vanished from context at
     every menu answer.
     """
-    loader.clear_skills_caches()
+    skills.clear_skills_caches()
     multi_step = (
         "analyzing-github-ci-performance",
         "scheduling-github-ci-fixes",
@@ -116,11 +116,11 @@ def test_multi_step_skills_track_progress_with_update_plan_not_step_headers() ->
         "delivering-morning-briefings",
     )
     for name in multi_step:
-        body = loader.load_skill_body(name)
+        body = skills.load_skill_body(name)
         assert "update_plan" in body, name
         assert "### [" not in body, name
     for name in (ONBOARDING_SKILL_NAME,):
-        assert "### [" not in loader.load_skill_body(name), name
+        assert "### [" not in skills.load_skill_body(name), name
 
 
 def test_capability_answers_and_direct_requests_do_not_require_onboarding() -> None:
@@ -161,7 +161,7 @@ def test_answer_keeps_skill_in_ephemeral_context_after_history_is_lost() -> None
     )
     answer = build_action_system_prompt_envelope(snapshot)
     fresh = build_action_system_prompt_envelope(replace(snapshot, text="Explain this deployment"))
-    body = loader.load_skill_body("analyzing-github-ci-performance")
+    body = skills.load_skill_body("analyzing-github-ci-performance")
     assert body in answer.render_ephemeral()
     assert body not in fresh.render()
     assert answer.render_cached() == fresh.render_cached()
@@ -182,19 +182,21 @@ def test_loader_discovers_nested_and_legacy_packages_without_hidden_directories(
         path = tmp_path / relative
         path.parent.mkdir(parents=True, exist_ok=True)
         path.write_text(skill_card(name, f"Body of {name}."))
-    monkeypatch.setattr(loader, "skills_dir", lambda: tmp_path)
-    loader.clear_skills_caches()
+    monkeypatch.setattr(
+        "core.agent_harness.prompts.skills.content.files.skills_dir", lambda: tmp_path
+    )
+    skills.clear_skills_caches()
     try:
-        assert [s.name for s in loader.list_action_skills()] == [
+        assert [s.name for s in skills.list_action_skills()] == [
             "master",
             "child",
             "legacy",
             "flat",
         ]
-        assert loader.load_skill_body("child") == "Body of child."
-        assert "master" in loader.load_skills_index()
+        assert skills.load_skill_body("child") == "Body of child."
+        assert "master" in skills.load_skills_index()
     finally:
-        loader.clear_skills_caches()
+        skills.clear_skills_caches()
 
 
 def test_includes_append_shared_markdown_once(
@@ -205,22 +207,24 @@ def test_includes_append_shared_markdown_once(
     (tmp_path / "demo.md").write_text(
         skill_card("demo", "Body of demo.", includes=["common/rule.md", "common/rule.md"])
     )
-    monkeypatch.setattr(loader, "skills_dir", lambda: tmp_path)
-    loader.clear_skills_caches()
+    monkeypatch.setattr(
+        "core.agent_harness.prompts.skills.content.files.skills_dir", lambda: tmp_path
+    )
+    skills.clear_skills_caches()
     try:
-        body = loader.load_skill_body("demo")
+        body = skills.load_skill_body("demo")
         assert body.startswith("Body of demo.")
         assert body.count("Do it once.") == 1
         assert "SHARED RULES from" in body
     finally:
-        loader.clear_skills_caches()
+        skills.clear_skills_caches()
 
 
 def test_onboarding_children_load_shared_rules_once() -> None:
-    loader.clear_skills_caches()
-    by_name = {s.name: s for s in loader.list_action_skills()}
-    reliability = loader.load_skill_body("scheduling-github-ci-fixes")
-    analytics = loader.load_skill_body("analyzing-github-ci-performance")
+    skills.clear_skills_caches()
+    by_name = {s.name: s for s in skills.list_action_skills()}
+    reliability = skills.load_skill_body("scheduling-github-ci-fixes")
+    analytics = skills.load_skill_body("analyzing-github-ci-performance")
     analytics_card = by_name["analyzing-github-ci-performance"].path.read_text(encoding="utf-8")
     # Shared rules resolve from the skills-tree ``common/`` folder and are
     # appended exactly once, never copied into the card body.
