+++ tests/core/agent_harness/session/test_pending_offer.py
@@ -457,7 +457,7 @@ def test_the_offer_tool_does_not_advertise_itself_as_the_way_to_run_a_report() -
 def test_the_skill_forbids_offering_before_the_work() -> None:
     """The recipe must state the ordering, not merely imply it by step number."""
     # Arrange
-    from core.agent_harness.prompts.skills.loader import skills_dir
+    from core.agent_harness.prompts.skills import skills_dir
 
     raw = (skills_dir() / "delivering-morning-briefings" / "SKILL.md").read_text(encoding="utf-8")
     body = " ".join(raw.replace("`", "").lower().split())
