+++ tests/scheduler/test_recurring_skill_acceptance.py
@@ -9,7 +9,7 @@ from click.testing import CliRunner
 
 import infrastructure.scheduling.scheduler.delivery_bundle as delivery_bundle
 from core.agent_harness import AgentSession, ToolCallingTurnResult, TurnResult
-from core.agent_harness.prompts.skills.loader import load_skill_body
+from core.agent_harness.prompts.skills import load_skill_body
 from infrastructure.scheduling.scheduler.executor import execute_task
 from infrastructure.scheduling.scheduler.storage.run_store import get_runs
 from infrastructure.scheduling.scheduler.storage.task_store import list_tasks
