+++ tests/scheduler/test_task_store.py
@@ -187,7 +187,7 @@ class TestStore:
 
 class TestRecurringSkillStoreIdentity:
     def test_changed_skill_revision_updates_existing_schedule(self, store_path: Path) -> None:
-        from core.agent_harness.prompts.skills.schedule import find_action_skill, skill_revision
+        from core.agent_harness.prompts.skills.scheduling import find_action_skill, skill_revision
 
         skill = find_action_skill("delivering-morning-briefings")
         assert skill is not None
